[Refactor] count 관련 필드의 자료형을 int -> long 으로 변경 #178

+++ src/main/java/com/yzgeneration/evc/domain/item/auctionitem/infrastructure/entity/AuctionItemStatsEntity.java
@@ -11,9 +11,9 @@ import lombok.*;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class AuctionItemStatsEntity {
 
-    private int viewCount;
+    private Long viewCount;
 
-    private int participantCount;
+    private Long participantCount;
 
     public static AuctionItemStatsEntity from(AuctionItemStats auctionItemStats){
         return AuctionItemStatsEntity.builder()
